[util/reggen] Addressed subword write.
If write size is smaller than the register width, then responses an
error. Previously it raises an error for subword write regardless of the
register width.

For instance, if the CPU writes a byte into a register having smaller
than 8 valid bit, it succeeds. But if register field width is greater
than 8, it returns error now.

Errored write access won't let the register value be written