Add 1 to the reqfifo depth
In the best case, sram can complete a transaction immediately on the next cycle.
When this happens, the adapter should immediately accept the next transaction.
However, if the reqfifo is sized to the same depth as the outstanding capacity,
it is not able to do so without creating a response -> request timing path.

As a result, an extra entry is added to the request fifo to handle this case.

The response fifo is not sized up as it can work as a passthrough fifo and
will not back-pressure on back to back responses unless the host is unwilling
to accept a response.