[topgen] Add PINMUX & PADS to hjson
This is a Pull-Request for defining PINMUX and PADS in hjson and
generating complete hjson struct to helping pinmux/ pads generators.
Purpose of this PR is to gather wide agreement what will come to the top
and the suggested data structure in `top_earlgrey.hjson` for PINMUX and
PADS.

This PR also removes unused hjson (deprecated) and clean-up the merge
script to make common functions to be used in validate, merge, and
inside the template.

This issue is related to #306