[util/vendor_hw.py] Add support for 'only_subdir' key in upstream section.
With an 'only_subdir' key in an upstream section, a single subtree of
an upstream repository can be specified for copying in.  This will
enable use of this tool to extract a small piece of a large upstream
repository without having to keep the path of directories down to that
piece and without having to maintain a large list of excludes.