[gpio] Revised to meet the style guide
Code was reviewed by @msfschaffner (#730) after GPIO signoff meeting.
Michael found a few stylish issues and one potential metastability
issue. For the metastability issue, it has been discussed in #45 and
decided to let it open.

The other two items are addressed in this PR. `_q` and `_d` naming rules
are both.