[ci] Perform Meson-driven software build on Ubuntu 16.04
We want to support 16.04 as a baseline requirement, so we should run our
CI on this.

Future improvements to CI might check the build on a larger matrix of
systems, but in the absence of that building on the oldest system we
hope to support seems the approach most likely to help in identifying
problems.