[build_docs] Add tools' selfdoc function
This brought back the functionality of `selfdoc` feature existed prior
to #638.

As each tool (`regtool`, `tlgen`) generates the selfdoc slightly
differently, current code manually generates case-by-case. It would be
beneficial if tool has common way to create selfdoc and gives the script
to detect whether it supports selfdoc feature or not.

Item 25 in #723 can be completed.