[dv/hmac] Fix TL outstanding request > 2
This is a fix for TL outstanding request >2. DV engineers decide to set
the max outstanding req to a very large value. Then when collecting
coverage, DV engineers will be able to provide a covered upper bound.
During coverage review, designers will be able to check if it meets their
expectation.
Signed-off-by: Cindy Chen <chencindy@google.com>