[reggen] Remove FIFO depth when window is used
Problem:

    reggen creates unintentional TL-UL buffers inside `tl_socket_1n` for
    window feature, which creates higher number of outstanding
    acceptances in the device port.

When `window` feature was implemented, `tlul_fifo_sync` doesn't have a
way to remove internal FIFO. So, when instantiating `tl_socket_1n` for
`window` requests, `ReqDepth` and `RspDepth` are set to 1h. It results
in the number of outstanding acceptances in a IP that having a window
port, to be more than or equal to 4 at least. One buffer in HReq,
another in DReq, 1 in DRsp, and 1 in DReq.

This PR is to remove the FIFO inside tlul_socket_1n, thanks to @tjaychen
added the condition of Depth == 0 and Pass == 1 case.

The IPs, for instance HMAC, still have a buffer internally. If
`tlul_adpater_sram` is used to convert TL-UL interface to the SRAM
interface, that module still has buffers internally.

The issue is discussed in #860

Signed-off-by: Eunchan Kim <eunchan@google.com>