[ci] Only run clang-format on C/C++ diffs
As noted in #920, the clang-format check in CI will attempt to reformat
.js files if present in the diff. This commit adds a whitelist, so only
diffs for filenames ending in .cpp, .cc, .c, or .h are passed to
clang-format-diff.

clang-format-diff (the current version on my machine at least) does have
options to filter paths on the input itself, but given we're already
filtering paths using grep it seemed better to consistently do filtering
in one place.

Signed-off-by: Alex Bradbury <asb@lowrisc.org>