[util] build_docs.py check if existing hugo is hugo-extended
build_docs.py uses hugo-extended to build the docs for OpenTitan, which
is stored as the binary named "hugo". When this binary is present the script
does not check whether the version is extended.

This change adds a hugo version check, which makes sure that the binary is
extended, and downloads the correct version if it is not.

Fixes #757
Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>