[reggen] Register to have typedef struct
In transition to the parameterized multireg #47, now each register has
its own struct packed type, which is named as
`module_reg2hw/hw2reg_regname_reg/mreg_t`. It creates mane typedef
inside the register package file but the overal reg2hw and hw2reg
structures are cleaner.

It gives more flexibility to transition to register interface from the
struct. The interface should consider the number of registers, not the
fields names.

There's still TODOs in this change.

1.  The field name function call `get_basename()` doesn't work properly
    with field having digits in its name. For instance, alert_handler
    has `en_e0` to `en_e3` in CLASSD_CTRL register, and the function
    chops the digit and returns `en_e` only, which creates an error.
2.  Handling the multi register and normal register is not standardized.
    The functions defined in `util/reggen/data.py` are specific for
    multireg type. It should be commonly used regardless of the register
    type.
3.  `util/topgen/lib.py` could be useful if it is moved to its own
    module. The functions inside lib.py can be used in any generator for
    RTLs.

This is related to #606 #270 #47

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>