[meson] Ensure we don't include a build-id in the elf
ld will sometimes (I guess it depends on how it's built) insert a
.note.gnu.build-id symbol into the ELF. This defaults to being at the
start of the ELF. If this happens we get non-code at the start of the
ELF which causes an illigal instruction to be executed when jumping from
ROM to a sw binary loaded into flash.

We could specify a location for the build-id (in the linker script) but
as it seems unessecay to include (it increases the size by 24 bytes) and
it isn't in the current generated release binaries it makes more sense
to just explicity disable it for all builds.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>