[sw] Standardise Variadic Logging Macros
In preparation for moving to `-std=c99`, we need to fix the usage of
variadic macros in the logging interface. This is because of a known
issue in C99/C11 where `__VA_ARGS__` often requires a non-zero number of
arguments in order to avoid issues with extra commas in function calls.

We fix the issue by including the initial, non-optional `fmt` argument
within the variadic part of these macros, so `__VA_ARGS__` always has
at least one argument within it.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>