[sw] Avoid using `__asm__` for Inline Assembly
In preparation for moving to `-std=c99`, all uses of `asm` as a C
keyword have become invalid.

Given we wish to use `asm`, this patch provides a define to `__asm__`, a
compiler-specific C extension which matches the `asm` keyword interface
in C++.

This patch also updates all uses of `__asm__` to use the `asm` macro,
following the notes in the proposed Assembly Style Guide.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>