[sw] Clean up the flash linker script and CRT.
This change brings the linker script and CRT to match with what's done
in #1213. Also, the CRT is now written in assembly, rather than C, since
having C code with an incomplete runtime (like missing interrupts or
uninitialized RAM sections) seems like a bad idea.

Given the similarity of flash_crt.S with its ROM counterpart, it may be
worth it to look into combining the two.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>