[doc] Change back to int for loop counters in example code
Our C and C++ style guide builds on top of Google's, which prefers int
for simple loop counters. There's relevant discussion on the advantages
of this approach on this llvm-dev thread, which also cites a range of
other sources:
* http://lists.llvm.org/pipermail/llvm-dev/2019-June/132890.html
* http://lists.llvm.org/pipermail/llvm-dev/2019-June/133023.html

PR #1240 changed two loops in example code from using int to using
size_t as a loop counter, and this commit reverts that change.

I'm posting this as a PR to test the waters to see if we're happy
sticking with our current approach or if someobody wants to champion a
proposed change to recommend size_t for loop counters.

Signed-off-by: Alex Bradbury <asb@lowrisc.org>