[topgen/tlgen] Move simplify_addr to tlgen
Previously, topgen simplified the address list of Xbar device port by
merging, compacting the multiple addresses into one.

The logic is self-contained, not affected by other crossbars. So it is
more logical that the tlgen has the optimization not in the topgen.

This commit is to refactor the code by moving the functions into tlgen.

Still it would be better that the optimization is in the elaboration
step not in the validation step. It was easier to move to validation as
the simplify function uses Hjson object not the Xbar instances.

Later, the logic can be tweaked by using Xbar instance (Node instances)
and do the optimization after validation step.

This closes #1302

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>